revised quest form
